+++ e2e-tests/hover_test.go
@@ -98,7 +98,7 @@ func TestHover(t *testing.T) {
 			result: &protocol.Hover{
 				Contents: protocol.MarkupContent{
 					Kind:  protocol.MarkupKindPlainText,
-					Value: "declared for backward compatibility, ignored.",
+					Value: "declared for backward compatibility, ignored. Please remove it.",
 				},
 			},
 		},